// dotnet/openapi/model/loginResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginResponse { 
    accessToken?: string | null;
    expiresIn?: number;
    refreshToken?: string | null;
    roles?: Array<string> | null;
    tokenType?: string | null;
    username?: string | null;
}